Add files via upload
My work so far for the stable marriage problem. So far the program reads a text document for each person's preferences and prints them to console, and then there's another method that finds the number of pairs of people so the program can be more flexible (ex. adding more people to the text file will not necessitate any change to the code).